+++ lib/banchan_web/live/auth_live/register_live.ex
@@ -15,8 +15,6 @@ defmodule BanchanWeb.RegisterLive do
 
   @impl true
   def mount(_params, session, socket) do
-    socket = assign_defaults(session, socket, false)
-
     {:ok,
      assign(socket,
        # For some reason, the tests for this aren't actually showing form
