+++ lib/banchan_web/live/home_live/index.ex
@@ -8,8 +8,7 @@ defmodule BanchanWeb.HomeLive do
   alias BanchanWeb.Components.{Layout, StudioCard}
 
   @impl true
-  def mount(_params, session, socket) do
-    socket = assign_defaults(session, socket, false)
+  def mount(_params, _session, socket) do
     studios = Studios.list_studios()
     {:ok, assign(socket, :studios, studios)}
   end
