--- lib/banchan_web/live/live_helpers.ex
@@ -1,33 +1,0 @@
-defmodule BanchanWeb.LiveHelpers do
-  @moduledoc false
-
-  import Phoenix.LiveView
-
-  alias Banchan.Accounts
-  alias Banchan.Accounts.User
-  alias BanchanWeb.Router.Helpers, as: Routes
-
-  def assign_defaults(session, socket, auth \\ true) do
-    # This is important so clients get booted when they log out elsewhere.
-    Accounts.subscribe_to_auth_events()
-
-    socket =
-      assign_new(socket, :current_user, fn ->
-        find_current_user(session)
-      end)
-
-    if auth && is_nil(socket.assigns.current_user) do
-      socket
-      |> put_flash(:error, "You must log in to access this page.")
-      |> redirect(to: Routes.user_session_path(socket, :create))
-    else
-      socket
-    end
-  end
-
-  defp find_current_user(session) do
-    with user_token when not is_nil(user_token) <- session["user_token"],
-         %User{} = user <- Accounts.get_user_by_session_token(user_token),
-         do: user
-  end
-end
