+++ lib/banchan_web/live/studio_live/commissions/new.ex
@@ -18,8 +18,7 @@ defmodule BanchanWeb.StudioLive.Commissions.New do
   import BanchanWeb.StudioLive.Helpers
 
   @impl true
-  def mount(%{"offering_type" => offering_type} = params, session, socket) do
-    socket = assign_defaults(session, socket, true)
+  def mount(%{"offering_type" => offering_type} = params, _session, socket) do
     socket = assign_studio_defaults(params, socket, false, true)
     offering = Offerings.get_offering_by_type!(offering_type, socket.assigns.current_user_member?)
