+++ lib/banchan_web/live/studio_live/index.ex
@@ -11,8 +11,7 @@ defmodule BanchanWeb.StudioLive.Index do
   alias BanchanWeb.Endpoint
 
   @impl true
-  def mount(_params, session, socket) do
-    socket = assign_defaults(session, socket, false)
+  def mount(_params, _session, socket) do
     studios = Studios.list_studios()
     {:ok, assign(socket, studios: studios)}
   end
