+++ lib/banchan_web/live/studio_live/offerings/new.ex
@@ -12,8 +12,7 @@ defmodule BanchanWeb.StudioLive.Offerings.New do
   import BanchanWeb.StudioLive.Helpers
 
   @impl true
-  def mount(params, session, socket) do
-    socket = assign_defaults(session, socket, true)
+  def mount(params, _session, socket) do
     socket = assign_studio_defaults(params, socket, true, false)
     changeset = Offering.changeset(%Offering{}, %{})
