+++ lib/banchan_web/live/studio_live/settings.ex
@@ -16,8 +16,7 @@ defmodule BanchanWeb.StudioLive.Settings do
   import BanchanWeb.StudioLive.Helpers
 
   @impl true
-  def mount(params, session, socket) do
-    socket = assign_defaults(session, socket, true)
+  def mount(params, _session, socket) do
     socket = assign_studio_defaults(params, socket, true, false)
     {:ok, assign(socket, changeset: Studio.changeset(socket.assigns.studio, %{}))}
   end
